refactored with a new hook for ddowns
